backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-generativeai==0.4.1
python-dotenv==1.0.1
cryptography==42.0.0
httpx==0.27.0
python-multipart==0.0.9
sentence-transformers==2.5.1
faiss-cpu==1.13.2
colorama==0.4.6
reportlab==4.1.0
pymupdf
